+++ css-selectors/src/components/game/task-complite.ts
