+++ css-selectors/src/components/game/task-forgame/task-06.ts
@@ -1,6 +1,6 @@
-import { CreateEltment } from '../classes/Create-element';
+import { CreateEltment } from '../../classes/Create-element';
 
-export function createLevel09(): void {
+export function createLevel06(): void {
   const elem1 = new CreateEltment('.current-task', 'div', 'car_3');
   const elem2 = new CreateEltment('.current-task', 'div', 'car_3');
   const elem3 = new CreateEltment('.current-task', 'div', 'train');
